Authorization With Policies

+++ app/Http/Controllers/ThreadsController.php
@@ -7,6 +7,7 @@ use App\Models\Thread;
 use App\Models\Channel;
 use Illuminate\Http\Request;
 use App\Filters\ThreadFilters;
+use Illuminate\Support\Facades\Redirect;
 
 class ThreadsController extends Controller
 {
@@ -113,6 +114,12 @@ class ThreadsController extends Controller
      */
     public function destroy($channel, Thread $thread)
     {
+        $this->authorize('update', $thread);
+        
+        if ($thread->user_id != auth()->id()) {
+            abort(403, 'You do not have permission to do this.');
+        }
+
         $thread->delete();
 
         if (request()->wantsJson()) {
